feat: implement Flask PCD backend with SocketIO and modular face blur integration
