+++ backend/services/__init__.py
@@ -1,0 +1,6 @@
+"""Services package for backend.
+
+This file makes the `services` directory a Python package so imports like
+`from services.pcd_main import process_frame_base64` work when running the
+backend application.
+"""
